implement A* using bucket queue, optimize data size.

+++ SPA/ASBQ.h
@@ -5,41 +5,43 @@
 #ifndef SPA_PICO_ASBQ_H
 #define SPA_PICO_ASBQ_H
 
+
 #include "../structure/BucketQueue.h"
 #include "SPA.h"
 
+
 class ASBQ : public SPA {
 private:
-    int maxWeight = 0;
+    short maxWeight = 0;
     Maze &maze;
 
     //The starting point and destination of route.
     Location *end;
     Location *start;
 
-    int maxRow, maxCol;
+    unsigned short maxRow, maxCol;
 
     // Distance table that stores shortest distance of all locations.
-    int **distTable;
+    short **distTable;
 
     // Priority Queue of adjacent locations to found set.
     // The int values indicate distance between current location and adjacent location +
     // estimated distance between adjacent location and destination.
-    BucketQueue<int, Location> *adjacentLocQueue;
+    BucketQueue<short, Location> *adjacentLocQueue;
 
     // Update distance of new adjacent location only.
     void UpdateDist(Location *currentLoc);
 
 public:
-    ASBQ(int maxRow, int maxCol, Maze &maze);
+    ASBQ(unsigned short maxRow, unsigned short maxCol, Maze &maze);
 
     ~ASBQ();
 
-    void setStart(int row, int column) override {
+    void setStart(unsigned short row, unsigned short column) override {
         start = &(maze.location[column][row]);
     }
 
-    void setEnd(int row, int column) override {
+    void setEnd(unsigned short row, unsigned short column) override {
         end = &(maze.location[column][row]);
     }
