+++ SPA/ASPQ.h
@@ -20,28 +20,28 @@ private:
     Location *start;
 
     Maze &maze;
-    int maxRow, maxCol;
+    unsigned short maxRow, maxCol;
 
     // Distance table that stores the shortest distance of all locations and their previous location.
-    int** distTable;
+    short** distTable;
 
     // Priority Queue of adjacent locations to found set.
     // The int values indicate distance between current location and adjacent location +
     // estimated distance between adjacent location and destination.
-    std::priority_queue<std::pair<int, Location *>> adjacentLocQueue;
+    std::priority_queue<std::pair<short, Location *>> adjacentLocQueue;
 
     // Update distance of new adjacent location only.
     void UpdateDist(Location *currentLoc);
 
 public:
-    ASPQ(int maxRow, int maxCol, Maze &maze);
+    ASPQ(unsigned short maxRow, unsigned short maxCol, Maze &maze);
     ~ASPQ();
 
-    void setStart(int row, int column) override {
+    void setStart(unsigned short row, unsigned short column) override {
         start = &(maze.location[column][row]);
     }
 
-    void setEnd(int row, int column) override {
+    void setEnd(unsigned short row, unsigned short column) override {
         end = &(maze.location[column][row]);
     }
